+++ ci/install_mykrobe_linux.sh
@@ -15,10 +15,11 @@ apt install -y \
   mongodb \
   python3-pip \
   python3-setuptools \
+  python-is-python3 \
   wget
 
 
-pip3 install tox
+python -m pip install -U pip
 
 cd $MYKROBE_ROOT_DIR
 rm -rf mccortex
@@ -26,18 +27,19 @@ git clone --recursive -b geno_kmer_count https://github.com/Mykrobe-tools/mccort
 cd mccortex
 make
 cd ..
+python -m pip install -r requirements.txt
+python -m pip install .
+# For whatever reason, mccortex is not getting put in the install location.
+# Do it manually.
+myk_dir=$(python -m pip show mykrobe | awk '/^Location/ {print $NF}')
+echo $myk_dir
+cp mccortex/bin/mccortex31 $myk_dir/mykrobe/cortex/mccortex31
 mkdir -p /data/db
 mongod --logpath mongo.log --quiet &>/dev/null &
 sleep 3s
-tox
+pytest --cov-report term-missing --cov=mykrobe
 mongod --shutdown
 rm mongo.log
-pip3 install .
-# For whatever reason, mccortex is not getting put in the install location.
-# Do it manually.
-myk_dir=$(pip3 show mykrobe | awk '/^Location/ {print $NF}')
-echo $myk_dir
-cp mccortex/bin/mccortex31 $myk_dir/mykrobe/cortex/mccortex31
 cd ..
 mykrobe panels update_metadata --debug
 mykrobe panels update_species --debug all